add utilities/ package; convert util/ to deprecation shims; move coco_classes to assets/
