+++ src/rfdetr/util/files.py
@@ -4,113 +4,20 @@
 # Licensed under the Apache License, Version 2.0 [see LICENSE for details]
 # ------------------------------------------------------------------------
 
-import hashlib
-import os
-import shutil
-from typing import Optional
+"""Deprecated: use ``rfdetr.utilities.files`` instead."""
 
-import requests
-from tqdm.auto import tqdm
+import warnings
 
-from rfdetr.util.logger import get_logger
+warnings.warn(
+    "rfdetr.util.files is deprecated; use rfdetr.utilities.files instead.",
+    DeprecationWarning,
+    stacklevel=2,
+)
 
-logger = get_logger()
+from rfdetr.utilities.files import (  # noqa: E402, F401
+    _compute_file_md5,
+    _download_file,
+    _validate_file_md5,
+)
 
-
-def _compute_file_md5(filepath: str) -> str:
-    """
-    Compute MD5 hash of a file.
-
-    Args:
-        filepath: Path to the file
-
-    Returns:
-        MD5 hash as hexadecimal string
-    """
-    md5_hash = hashlib.md5()
-    with open(filepath, "rb") as f:
-        for chunk in iter(lambda: f.read(4096), b""):
-            md5_hash.update(chunk)
-    return md5_hash.hexdigest()
-
-
-def _validate_file_md5(filepath: str, expected_md5: str) -> bool:
-    """
-    Validate that a file's MD5 hash matches the expected hash.
-
-    Args:
-        filepath: Path to the file
-        expected_md5: Expected MD5 hash
-
-    Returns:
-        True if hash matches, False otherwise
-    """
-    if not os.path.exists(filepath):
-        return False
-
-    actual_md5 = _compute_file_md5(filepath)
-    return actual_md5.lower() == expected_md5.lower()
-
-
-def _download_file(url: str, filename: str, expected_md5: Optional[str] = None) -> None:
-    """
-    Download a file from a URL with optional MD5 validation.
-
-    Args:
-        url: URL to download from
-        filename: Path to save the file
-        expected_md5: Expected MD5 hash for validation (optional)
-
-    Raises:
-        ValueError: If MD5 validation fails
-    """
-    # Check if file exists and has correct hash
-    if os.path.exists(filename) and expected_md5:
-        if _validate_file_md5(filename, expected_md5):
-            logger.info(f"File {filename} already exists with correct MD5 hash. Skipping download.")
-            return
-        else:
-            logger.warning(f"File {filename} exists but MD5 hash mismatch. Re-downloading...")
-            os.remove(filename)
-
-    response = requests.get(url, stream=True)
-    response.raise_for_status()
-    total_size_header = response.headers.get("content-length")
-    try:
-        total_size = int(total_size_header) if total_size_header is not None else None
-    except (TypeError, ValueError):
-        total_size = None
-
-    # Download to temporary file first
-    temp_filename = f"{filename}.tmp"
-    try:
-        with (
-            open(temp_filename, "wb") as f,
-            tqdm(
-                desc=filename,
-                total=total_size,
-                unit="iB",
-                unit_scale=True,
-                unit_divisor=1024,
-            ) as pbar,
-        ):
-            for data in response.iter_content(chunk_size=1024):
-                size = f.write(data)
-                pbar.update(size)
-    except Exception:
-        if os.path.exists(temp_filename):
-            os.remove(temp_filename)
-        raise
-
-    # Validate MD5 if expected hash is provided.
-    if expected_md5:
-        actual_md5 = _compute_file_md5(temp_filename)
-        if actual_md5.lower() != expected_md5.lower():
-            if os.path.exists(temp_filename):
-                os.remove(temp_filename)
-            raise ValueError("MD5 mismatch for %s (expected %s, got %s)." % (filename, expected_md5, actual_md5))
-        else:
-            logger.info(f"MD5 validation successful for {filename}")
-
-    # shutil.move handles cross-device moves (e.g. tmpfs → ext4 on Colab).
-    shutil.move(temp_filename, filename)
+__all__ = ["_compute_file_md5", "_download_file", "_validate_file_md5"]
