+++ src/rfdetr/util/misc.py
@@ -15,254 +15,59 @@
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
 # ------------------------------------------------------------------------
 
-"""
-Misc functions, including distributed helpers.
+"""Deprecated: most symbols have moved to ``rfdetr.utilities``.
 
-Mostly copy-paste from torchvision references.
+``accuracy``, ``inverse_sigmoid``, and ``interpolate`` remain here temporarily
+and will move to ``rfdetr.models.math`` in Phase 10.
 """
 
-import os
-import pickle
-import subprocess
-import tempfile
-from typing import Any, Dict, List, Optional, Tuple
+import warnings
+
+warnings.warn(
+    "rfdetr.util.misc is deprecated; use rfdetr.utilities instead.",
+    DeprecationWarning,
+    stacklevel=2,
+)
+
+# Re-export symbols that have moved to utilities/.
+from rfdetr.utilities.distributed import (  # noqa: E402, F401
+    all_gather,
+    get_rank,
+    get_world_size,
+    is_dist_avail_and_initialized,
+    is_main_process,
+    reduce_dict,
+    save_on_master,
+)
+from rfdetr.utilities.package import get_sha  # noqa: E402, F401
+from rfdetr.utilities.state_dict import strip_checkpoint  # noqa: E402, F401
+from rfdetr.utilities.tensors import (  # noqa: E402, F401
+    NestedTensor,
+    _max_by_axis,
+    _onnx_nested_tensor_from_tensor_list,
+    collate_fn,
+    nested_tensor_from_tensor_list,
+)
+
+# ---------------------------------------------------------------------------
+# accuracy, inverse_sigmoid, interpolate remain here until Phase 10
+# when they move to rfdetr.models.math.
+# ---------------------------------------------------------------------------
+
+from typing import List, Optional, Tuple
 
 import torch
-import torch.distributed as dist
-
-# needed due to empty tensor bug in pytorch and torchvision 0.5
 import torchvision
 from torch import Tensor
-
-from rfdetr.util.logger import get_logger
-
-logger = get_logger()
 
 if float(torchvision.__version__.split(".")[1]) < 7.0:
     from torchvision.ops import _new_empty_tensor
     from torchvision.ops.misc import _output_size
 
 
-def all_gather(data: Any) -> List[Any]:
-    """
-    Run all_gather on arbitrary picklable data (not necessarily tensors)
-    Args:
-        data: any picklable object
-    Returns:
-        list of data gathered from each rank
-    """
-    world_size = get_world_size()
-    if world_size == 1:
-        return [data]
-
-    # serialized to a Tensor
-    buffer = pickle.dumps(data)
-    storage = torch.ByteStorage.from_buffer(buffer)
-    tensor = torch.ByteTensor(storage).to("cuda")
-
-    # obtain Tensor size of each rank
-    local_size = torch.tensor([tensor.numel()], device="cuda")
-    size_list = [torch.tensor([0], device="cuda") for _ in range(world_size)]
-    dist.all_gather(size_list, local_size)
-    size_list = [int(size.item()) for size in size_list]
-    max_size = max(size_list)
-
-    # receiving Tensor from all ranks
-    # we pad the tensor because torch all_gather does not support
-    # gathering tensors of different shapes
-    tensor_list = []
-    for _ in size_list:
-        tensor_list.append(torch.empty((max_size,), dtype=torch.uint8, device="cuda"))
-    if local_size != max_size:
-        padding = torch.empty(size=(max_size - local_size,), dtype=torch.uint8, device="cuda")
-        tensor = torch.cat((tensor, padding), dim=0)
-    dist.all_gather(tensor_list, tensor)
-
-    data_list = []
-    for size, tensor in zip(size_list, tensor_list):
-        buffer = tensor.cpu().numpy().tobytes()[:size]
-        data_list.append(pickle.loads(buffer))
-
-    return data_list
-
-
-def reduce_dict(input_dict: Dict[str, torch.Tensor], average: bool = True) -> Dict[str, torch.Tensor]:
-    """
-    Args:
-        input_dict (dict): all the values will be reduced
-        average (bool): whether to do average or sum
-    Reduce the values in the dictionary from all processes so that all processes
-    have the averaged results. Returns a dict with the same fields as
-    input_dict, after reduction.
-    """
-    world_size = get_world_size()
-    if world_size < 2:
-        return input_dict
-    with torch.no_grad():
-        names = []
-        values = []
-        # sort the keys so that they are consistent across processes
-        for k in sorted(input_dict.keys()):
-            names.append(k)
-            values.append(input_dict[k])
-        values = torch.stack(values, dim=0)
-        dist.all_reduce(values)
-        if average:
-            values /= world_size
-        reduced_dict = {k: v for k, v in zip(names, values)}
-    return reduced_dict
-
-
-def get_sha() -> str:
-    """Return a short status string for the current git repo, or 'unknown' if unavailable."""
-    cwd = os.path.dirname(os.path.abspath(__file__))
-
-    def _run(command: List[str]) -> str:
-        return subprocess.check_output(command, cwd=cwd).decode("ascii").strip()
-
-    try:
-        sha = _run(["git", "rev-parse", "HEAD"])
-        has_diff = bool(_run(["git", "diff-index", "HEAD"]))
-        status = "has uncommitted changes" if has_diff else "clean"
-        branch = _run(["git", "rev-parse", "--abbrev-ref", "HEAD"])
-        return f"sha: {sha}, status: {status}, branch: {branch}"
-    except (subprocess.CalledProcessError, FileNotFoundError):
-        return "unknown"
-
-
-def collate_fn(batch: List[Tuple[Any, ...]]) -> Tuple[Any, ...]:
-    batch = list(zip(*batch))
-    batch[0] = nested_tensor_from_tensor_list(batch[0])
-    return tuple(batch)
-
-
-def _max_by_axis(the_list: List[List[int]]) -> List[int]:
-    maxes = the_list[0]
-    for sublist in the_list[1:]:
-        for index, item in enumerate(sublist):
-            maxes[index] = max(maxes[index], item)
-    return maxes
-
-
-class NestedTensor(object):
-    def __init__(self, tensors: Tensor, mask: Optional[Tensor]) -> None:
-        self.tensors = tensors
-        self.mask = mask
-
-    def to(self, device: torch.device, **kwargs: Any) -> "NestedTensor":
-        cast_tensor = self.tensors.to(device, **kwargs)
-        mask = self.mask
-        if mask is not None:
-            assert mask is not None
-            cast_mask = mask.to(device, **kwargs)
-        else:
-            cast_mask = None
-        return NestedTensor(cast_tensor, cast_mask)
-
-    def pin_memory(self) -> "NestedTensor":
-        return NestedTensor(
-            self.tensors.pin_memory(),
-            self.mask.pin_memory() if self.mask is not None else None,
-        )
-
-    def decompose(self) -> Tuple[Tensor, Optional[Tensor]]:
-        return self.tensors, self.mask
-
-    def __repr__(self) -> str:
-        return str(self.tensors)
-
-
-def nested_tensor_from_tensor_list(tensor_list: List[Tensor]) -> NestedTensor:
-    # TODO make this more general
-    if tensor_list[0].ndim == 3:
-        if torchvision._is_tracing():
-            # nested_tensor_from_tensor_list() does not export well to ONNX
-            # call _onnx_nested_tensor_from_tensor_list() instead
-            return _onnx_nested_tensor_from_tensor_list(tensor_list)
-
-        # TODO make it support different-sized images
-        max_size = _max_by_axis([list(img.shape) for img in tensor_list])
-        # min_size = tuple(min(s) for s in zip(*[img.shape for img in tensor_list]))
-        batch_shape = [len(tensor_list)] + max_size
-        b, c, h, w = batch_shape
-        dtype = tensor_list[0].dtype
-        device = tensor_list[0].device
-        tensor = torch.zeros(batch_shape, dtype=dtype, device=device)
-        mask = torch.ones((b, h, w), dtype=torch.bool, device=device)
-        for img, pad_img, m in zip(tensor_list, tensor, mask):
-            pad_img[: img.shape[0], : img.shape[1], : img.shape[2]].copy_(img)
-            m[: img.shape[1], : img.shape[2]] = False
-    else:
-        raise ValueError("not supported")
-    return NestedTensor(tensor, mask)
-
-
-# _onnx_nested_tensor_from_tensor_list() is an implementation of
-# nested_tensor_from_tensor_list() that is supported by ONNX tracing.
-@torch.jit.unused
-def _onnx_nested_tensor_from_tensor_list(tensor_list: List[Tensor]) -> NestedTensor:
-    max_size = []
-    for i in range(tensor_list[0].dim()):
-        max_size_i = torch.max(torch.stack([img.shape[i] for img in tensor_list]).to(torch.float32)).to(torch.int64)
-        max_size.append(max_size_i)
-    max_size = tuple(max_size)
-
-    # work around for
-    # pad_img[: img.shape[0], : img.shape[1], : img.shape[2]].copy_(img)
-    # m[: img.shape[1], :img.shape[2]] = False
-    # which is not yet supported in onnx
-    padded_imgs = []
-    padded_masks = []
-    for img in tensor_list:
-        padding = [(s1 - s2) for s1, s2 in zip(max_size, tuple(img.shape))]
-        padded_img = torch.nn.functional.pad(img, (0, padding[2], 0, padding[1], 0, padding[0]))
-        padded_imgs.append(padded_img)
-
-        m = torch.zeros_like(img[0], dtype=torch.int, device=img.device)
-        padded_mask = torch.nn.functional.pad(m, (0, padding[2], 0, padding[1]), "constant", 1)
-        padded_masks.append(padded_mask.to(torch.bool))
-
-    tensor = torch.stack(padded_imgs)
-    mask = torch.stack(padded_masks)
-
-    return NestedTensor(tensor, mask=mask)
-
-
-def is_dist_avail_and_initialized():
-    if not dist.is_available():
-        return False
-    if not dist.is_initialized():
-        return False
-    return True
-
-
-def get_world_size():
-    if not is_dist_avail_and_initialized():
-        return 1
-    return dist.get_world_size()
-
-
-def get_rank():
-    if not is_dist_avail_and_initialized():
-        return 0
-    return dist.get_rank()
-
-
-def is_main_process():
-    return get_rank() == 0
-
-
-def save_on_master(obj, f, *args, **kwargs):
-    """
-    Safely save objects, removing any callbacks that can't be pickled
-    """
-    if is_main_process():
-        torch.save(obj, f, *args, **kwargs)
-
-
 @torch.no_grad()
 def accuracy(output: torch.Tensor, target: torch.Tensor, topk: Tuple[int, ...] = (1,)) -> List[torch.Tensor]:
-    """Computes the precision@k for the specified values of k"""
+    """Computes the precision@k for the specified values of k."""
     if target.numel() == 0:
         return [torch.zeros([], device=output.device)]
     maxk = max(topk)
@@ -286,11 +91,7 @@ def interpolate(
     mode: str = "nearest",
     align_corners: Optional[bool] = None,
 ) -> Tensor:
-    """
-    Equivalent to nn.functional.interpolate, but with support for empty batch sizes.
-    This will eventually be supported natively by PyTorch, and this
-    class can go away.
-    """
+    """Equivalent to nn.functional.interpolate, but with support for empty batch sizes."""
     if float(torchvision.__version__.split(".")[1]) < 7.0:
         if input.numel() > 0:
             return torch.nn.functional.interpolate(input, size, scale_factor, mode, align_corners)
@@ -309,20 +110,27 @@ def inverse_sigmoid(x: torch.Tensor, eps: float = 1e-5) -> torch.Tensor:
     return torch.log(x1 / x2)
 
 
-def strip_checkpoint(checkpoint: str | os.PathLike[str]) -> None:
-    state_dict = torch.load(checkpoint, map_location="cpu", weights_only=False)
-    new_state_dict = {
-        "model": state_dict["model"],
-        "args": state_dict["args"],
-    }
-    # Create the temp file in the destination directory so os.replace stays on the same filesystem (atomic).
-    checkpoint_dir = os.path.dirname(os.path.abspath(os.fspath(checkpoint)))
-    with tempfile.NamedTemporaryFile(dir=checkpoint_dir, delete=False) as tmp_file:
-        tmp_path = tmp_file.name
-    try:
-        torch.save(new_state_dict, tmp_path)
-        # Atomic replace avoids leaving a partially written checkpoint on save failures/interruption.
-        os.replace(tmp_path, checkpoint)
-    finally:
-        if os.path.exists(tmp_path):
-            os.remove(tmp_path)
+__all__ = [
+    # distributed
+    "all_gather",
+    "get_rank",
+    "get_world_size",
+    "is_dist_avail_and_initialized",
+    "is_main_process",
+    "reduce_dict",
+    "save_on_master",
+    # tensors
+    "NestedTensor",
+    "collate_fn",
+    "nested_tensor_from_tensor_list",
+    "_max_by_axis",
+    "_onnx_nested_tensor_from_tensor_list",
+    # package
+    "get_sha",
+    # state_dict
+    "strip_checkpoint",
+    # math (temporarily here, move to models.math in Phase 10)
+    "accuracy",
+    "interpolate",
+    "inverse_sigmoid",
+]
