+++ tests/benchmarks/conftest.py
@@ -12,7 +12,7 @@ from rfdetr.datasets._develop import (
     _download_and_extract,
     _download_lock,
 )
-from rfdetr.util.utils import seed_all
+from rfdetr.utilities.reproducibility import seed_all
 
 _PROJECT_ROOT = Path(__file__).resolve().parents[2]
 _DATA_DIR = _PROJECT_ROOT / "data"
